added options for transpilation
