+++ pancake2viper/src/translation/mod.rs
@@ -1,5 +1,7 @@
+pub mod context;
 pub mod expression;
 pub mod statement;
 pub mod top;
 
-pub use top::{ToViper, ViperEncodeCtx};
+pub use context::ViperEncodeCtx;
+pub use top::ToViper;
